Introduced the generic RegisterInfo and RegisterCatalogue classes to provide abstract register information for all backends. The new plugin mechanism works with these new classes for any backends (given that the plugins can be configured in the backend's map file).



git-svn-id: https://svnsrv.desy.de/desy/mtca4u/deviceaccess/trunk@2458 5f601a5c-88b9-47f3-b851-818f6741d00d

+++ device/src/FixedPointConverter.cc
@@ -79,7 +79,7 @@ namespace mtca4u {
         return toRaw(std::stoi(cookedValue));
       }
       else {
-        return toRaw(std::stoul(cookedValue));
+        return toRaw(static_cast<uint32_t>(std::stoul(cookedValue)));   // on some compilers, long might be a different type than int...
       }
     }
     else {
